refactor(runtime): split work item tests into focused modules
